Add MIT License and update REUSE compliance
Introduced the MIT License by adding a `LICENSE` file to the project. Updated the `REUSE.toml` file to include the `LICENSE` file in the `path` field under `[[annotations]]`. Set the `SPDX-License-Identifier` to `MIT` and updated the `SPDX-FileCopyrightText` to reflect the year 2024 and copyright ownership by "Carl Zeiss Microscopy GmbH." Ensured SPDX compliance and proper documentation of licensing terms.